Moves IPlayer to 'model' package
